backend/requirements.txt
Django==3.2.8
psycopg2-binary==2.9
django-cors-headers==3.11
djangorestframework==3.13
djangorestframework-simplejwt==5.1
drf-yasg==1.20
gunicorn==20.1
dj-database-url==0.5
django-heroku==0.3
whitenoise==5.3
django-filter==2.4.0
celery[redis]==4.4.2
